split plugins into individual files
