+++ .config/nvim/lua/plugins/diffview.lua
@@ -1,0 +1,4 @@
+-- diffview provides an interface for exploring git history
+return {
+  { "sindrets/diffview.nvim", cmd = { "DiffviewOpen", "DiffviewFileHistory" } },
+}
