+++ .config/nvim/lua/plugins/gv.lua
@@ -1,0 +1,8 @@
+-- gv.vim provides a graphical git log
+return {
+  {
+    "junegunn/gv.vim",
+    cmd = "GV",
+    dependencies = { "tpope/vim-fugitive" }
+  },
+}
