+++ .config/nvim/lua/plugins/indent-blankline.lua
@@ -1,0 +1,13 @@
+-- indent-blankline shows vertical lines at tabstops
+return {
+  {
+    "lukas-reineke/indent-blankline.nvim",
+    lazy = false,
+    main = "ibl",
+    opts = {
+      -- make the character a little less dense
+      indent = { char = "┊" },
+      exclude = { filetypes = { "markdown", "rst" } },
+    },
+  },
+}
