+++ .config/nvim/lua/plugins/lsp-progress.lua
@@ -1,0 +1,7 @@
+-- lsp-progress adds LSP status to the bottom line so you know it's running
+return {
+  {
+    "linrongbin16/lsp-progress.nvim",
+    config = true,
+  },
+}
