+++ .config/nvim/lua/plugins/nightfox.lua
@@ -1,0 +1,11 @@
+-- nightfox is a family of colorschemes (nightfox, dawnfox, terrafox, etc)
+return {
+  {
+    "EdenEast/nightfox.nvim",
+    lazy = false,
+    priority = 1000,
+    opts = {
+      styles = { comments = "italic" },
+    },
+  },
+}
