+++ .config/nvim/lua/plugins/nvim-colorizer.lua
@@ -1,0 +1,9 @@
+-- nvim-colorizer colors hex codes by their actual color
+return {
+  {
+    "norcalli/nvim-colorizer.lua",
+    config = true,
+    lazy = true,
+    cmd = "ColorizerToggle"
+  }, 
+}
