+++ .config/nvim/lua/plugins/vim-commentary.lua
@@ -1,0 +1,4 @@
+-- vim-commentary toggles comments
+return {
+  { "tpope/vim-commentary" },
+}
