+++ .config/nvim/lua/plugins/vim-diff-enhanced.lua
@@ -1,0 +1,4 @@
+-- vim-diff-enhanced provides other diff algorithms
+return {
+  { "chrisbra/vim-diff-enhanced" },
+}
