+++ .config/nvim/lua/plugins/vim-fugitive.lua
@@ -1,0 +1,4 @@
+ -- vim-fugitive provides a convenient git interface, with incremental commits
+ return {
+  { "tpope/vim-fugitive", cmd = "Git", lazy = true },
+ }
