+++ .config/nvim/lua/plugins/vim-mergetool.lua
@@ -1,0 +1,7 @@
+-- vim-mergetool lets you easily work with 3-way merge conflicts
+return {
+  {
+    "samoshkin/vim-mergetool",
+    cmd = "MergetoolStart",
+  },
+}
