+++ .config/nvim/lua/plugins/vim-python-pep8-indent.lua
@@ -1,0 +1,7 @@
+-- vim-python-pep8-indent gives better indentation for python
+return {
+  {
+    "daler/vim-python-pep8-indent",
+    ft = { "python", "snakemake" },
+  },
+}
