+++ .config/nvim/lua/plugins/vim-sleuth.lua
@@ -1,0 +1,4 @@
+-- vim-sleuth autodetects indentation settings
+return {
+  { "tpope/vim-sleuth" },
+}
