+++ .config/nvim/lua/plugins/vim-surround.lua
@@ -1,0 +1,4 @@
+-- vim-surround provides tools for easily changing surrounding characters
+return {
+  { "tpope/vim-surround" },
+}
