+++ .config/nvim/lua/plugins/vim-table-mode.lua
@@ -1,0 +1,4 @@
+-- vim-table-mode lets you easily work with ReST and Markdown tables
+return {
+  { "dhruvasagar/vim-table-mode" },
+}
