+++ .config/nvim/lua/plugins/vim-tmux-clipboard.lua
@@ -1,0 +1,4 @@
+-- vim-tmux-clipboard makes the cliboard work better with tmux
+return {
+  { "roxma/vim-tmux-clipboard" },
+}
