+++ .config/nvim/lua/plugins/vis.lua
@@ -1,0 +1,4 @@
+-- vis restricts replacement only to visual selection
+return {
+  { "vim-scripts/vis" },
+}
