+++ .config/nvim/lua/plugins/which-key.lua
@@ -1,0 +1,13 @@
+-- which-key pops up a window showing possible keybindings
+return {
+  {
+    "folke/which-key.nvim",
+
+    -- later versions raise errors on conflicting keybindings
+    commit = "0539da005b98b02cf730c1d9da82b8e8edb1c2d2",
+
+    lazy = false,
+    config = true,
+
+  },
+}
